apply mvvm structure
apply mvvm structure
using (DataBinding, LiveData, AAC ViewModel)

1. separate package structure
2. add MainViewModel
3. add MainBindinAdapater
