+++ app/build.gradle
@@ -48,15 +48,16 @@ android {
 dependencies {
 
     def work_version = "2.4.0"
+    def lifecycle_version = "2.2.0"
 
     implementation fileTree(dir: 'libs', include: ['*.jar'])
     implementation 'com.google.guava:listenablefuture:9999.0-empty-to-avoid-conflict-with-guava'
     implementation "org.jetbrains.kotlin:kotlin-stdlib-jdk7:$kotlin_version"
     implementation 'androidx.appcompat:appcompat:1.2.0'
-    implementation 'androidx.core:core-ktx:1.3.1'
+    implementation 'androidx.core:core-ktx:1.3.2'
 
     implementation 'com.google.firebase:firebase-analytics:17.5.0'
-    implementation 'com.google.firebase:firebase-messaging:20.2.4'
+    implementation 'com.google.firebase:firebase-messaging:20.3.0'
     kapt 'com.android.databinding:compiler:4.0.1'
 
     implementation "androidx.work:work-runtime-ktx:$work_version"
@@ -69,21 +70,21 @@ dependencies {
 
     implementation 'androidx.recyclerview:recyclerview:1.1.0'
     implementation 'androidx.cardview:cardview:1.0.0'
-    implementation 'com.google.android.material:material:1.2.0'
+    implementation 'com.google.android.material:material:1.2.1'
     implementation 'com.github.bumptech.glide:glide:4.11.0'
 
-    implementation "org.jetbrains.kotlinx:kotlinx-coroutines-android:1.3.8"
+    implementation "org.jetbrains.kotlinx:kotlinx-coroutines-android:1.3.9"
     implementation "androidx.swiperefreshlayout:swiperefreshlayout:1.1.0"
     kapt 'com.github.bumptech.glide:compiler:4.11.0'
 
     implementation 'androidx.legacy:legacy-support-v4:1.0.0' // DrawerLayout
-    implementation 'com.google.android.material:material:1.2.0' // NavigationView
+    implementation 'com.google.android.material:material:1.2.1' // NavigationView
 
-    implementation 'org.jsoup:jsoup:1.13.1' //jsoup
-    implementation group: 'org.seleniumhq.selenium', name: 'selenium-java', version: '3.141.59'
-    //selenium
+
+    implementation group: 'androidx.lifecycle', name:'lifecycle-viewmodel-ktx', version: '2.3.0-beta01'
+    implementation "androidx.lifecycle:lifecycle-livedata-ktx:$lifecycle_version"
 
     testImplementation 'junit:junit:4.13'
-    androidTestImplementation 'androidx.test.ext:junit:1.1.1'
-    androidTestImplementation 'androidx.test.espresso:espresso-core:3.2.0'
+    androidTestImplementation 'androidx.test.ext:junit:1.1.2'
+    androidTestImplementation 'androidx.test.espresso:espresso-core:3.3.0'
 }
