+++ app/src/main/java/com/example/lol_notification_project/data/model/League.kt
@@ -1,4 +1,4 @@
-package com.example.lol_notification_project.model.data
+package com.example.lol_notification_project.data.model
 
 class LeagueEntryDTO {
     var leagueId: String? = null
