+++ app/src/main/res/layout/activity_main.xml
@@ -8,7 +8,7 @@
     <data>
         <variable
             name="vm"
-            type="com.example.lol_notification_project.viewmodel.CardViewModel" />
+            type="com.example.lol_notification_project.ui.main.MainViewModel" />
     </data>
 
     <androidx.drawerlayout.widget.DrawerLayout
@@ -20,7 +20,7 @@
         <androidx.constraintlayout.widget.ConstraintLayout
             android:layout_width="match_parent"
             android:layout_height="match_parent"
-            tools:context=".view.MainActivity">
+            tools:context=".ui.main.MainActivity">
 
             <include layout="@layout/main_toolbar" />
 
@@ -29,6 +29,7 @@
                 android:layout_width="match_parent"
                 android:layout_height="match_parent"
                 android:layout_marginTop="?attr/actionBarSize"
+                app:refreshing="@{vm.swipe_refresh}"
                 app:layout_constraintEnd_toEndOf="parent"
                 app:layout_constraintStart_toStartOf="parent"
                 app:layout_constraintTop_toTopOf="parent">
@@ -43,23 +44,9 @@
                     app:layout_constraintHorizontal_bias="0.0"
                     app:layout_constraintStart_toStartOf="parent"
                     app:layout_constraintTop_toTopOf="parent"
-                    app:viewModel="@{vm}"
                     app:summoner="@{vm.summonerInfo}" />
 
             </androidx.swiperefreshlayout.widget.SwipeRefreshLayout>
-
-            <ProgressBar
-                android:id="@+id/progressBar"
-                style="?android:attr/progressBarStyle"
-                android:progressBackgroundTint="@color/colorPrimary"
-                android:progressTint="@color/colorPrimary"
-                android:layout_width="100dp"
-                android:layout_height="100dp"
-                app:layout_constraintBottom_toBottomOf="parent"
-                app:layout_constraintEnd_toEndOf="parent"
-                app:layout_constraintStart_toStartOf="parent"
-                app:layout_constraintTop_toTopOf="parent" />
-
         </androidx.constraintlayout.widget.ConstraintLayout>
 
         <com.google.android.material.navigation.NavigationView
@@ -71,7 +58,6 @@
             app:menu="@menu/drawer_navigation">
 
         </com.google.android.material.navigation.NavigationView>
-
 
     </androidx.drawerlayout.widget.DrawerLayout>
 </layout>
