+++ app/src/main/res/layout/activity_web_view.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".view.webViewActivity">
+    tools:context=".ui.webview.webViewActivity">
 
     <WebView
     android:id="@+id/webView"
